Move transaction-sync-only code out of internal/activities; use real durations for batch flush
internal/activities is meant for Temporal-registered activities; the
transaction claim/fetch/flush code and its types had no Temporal
caller left after the previous commit retired the Temporal path, so
they move to internal/transactioncron (their only consumer) as plain
functions. IsPlayerOnlyTransaction stays in activities, exported,
since ScavengerActivities' replication also depends on it.

CRON_TXN_BATCH_FLUSH_INTERVAL_SECONDS becomes
CRON_TXN_BATCH_FLUSH_INTERVAL_DURATION, parsed as a time.Duration
(helpers.GetEnv gains a time.Duration case via time.ParseDuration)
instead of a plain int-seconds count.

Also drops a few comments that narrated the now-completed Temporal
retirement instead of describing current behavior.
