+++ backend/internal/activities/claim_pg_test.go
@@ -45,115 +45,6 @@ func seedLeague(t *testing.T, db *gorm.DB, l models.SleeperLeague) {
 	}
 }
 
-func TestClaimLeagues_OrderingLimitAndStamp(t *testing.T) {
-	db := newPGTestDB(t)
-	now := time.Now().UTC()
-	old := now.Add(-48 * time.Hour)
-	recent := now.Add(-1 * time.Hour)
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "never", LastFetchedAt: &now})
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "oldest", LastFetchedAt: &now, LastTransactionsFetchedAt: &old})
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "recent", LastFetchedAt: &now, LastTransactionsFetchedAt: &recent})
-
-	a := &activities.DataFetchActivities{DB: db}
-	got, err := a.ClaimLeaguesForTransactions(context.Background(), db, activities.ClaimLeaguesForTransactionsParams{BatchSize: 2})
-	if err != nil {
-		t.Fatalf("claim: %v", err)
-	}
-	// Which leagues get claimed is what matters (RETURNING order is not
-	// guaranteed): batch of 2 must be the never-fetched and the oldest.
-	claimed := map[string]bool{}
-	for _, s := range got {
-		claimed[s.LeagueID] = true
-		if s.Season != "2026" {
-			t.Errorf("expected Season populated, got %+v", s)
-		}
-	}
-	if len(got) != 2 || !claimed["never"] || !claimed["oldest"] {
-		t.Fatalf("expected {never, oldest}, got %+v", got)
-	}
-	var stamped int64
-	db.Model(&models.SleeperLeague{}).Where("claimed_at IS NOT NULL").Count(&stamped)
-	if stamped != 2 {
-		t.Errorf("expected 2 rows stamped claimed_at, got %d", stamped)
-	}
-}
-
-func TestClaimLeagues_ExcludesIneligible(t *testing.T) {
-	db := newPGTestDB(t)
-	now := time.Now().UTC()
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "skipped", LastFetchedAt: &now, SkippedAt: &now})
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "unfetched"}) // last_fetched_at NULL
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "old-season", Season: "2024", LastFetchedAt: &now})
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "done-complete", Status: "complete", LastFetchedAt: &now, LastTransactionsFetchedAt: &now})
-	// complete but never transaction-synced: still eligible
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "complete-unsynced", Status: "complete", LastFetchedAt: &now})
-
-	a := &activities.DataFetchActivities{DB: db}
-	got, err := a.ClaimLeaguesForTransactions(context.Background(), db, activities.ClaimLeaguesForTransactionsParams{BatchSize: 10})
-	if err != nil {
-		t.Fatalf("claim: %v", err)
-	}
-	if len(got) != 1 || got[0].LeagueID != "complete-unsynced" {
-		t.Fatalf("expected only complete-unsynced, got %+v", got)
-	}
-}
-
-func TestClaimLeagues_RespectsAndExpiresClaims(t *testing.T) {
-	db := newPGTestDB(t)
-	now := time.Now().UTC()
-	fresh := now.Add(-1 * time.Minute)
-	stale := now.Add(-30 * time.Minute)
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "fresh-claim", LastFetchedAt: &now, ClaimedAt: &fresh})
-	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "expired-claim", LastFetchedAt: &now, ClaimedAt: &stale})
-
-	a := &activities.DataFetchActivities{DB: db}
-	got, err := a.ClaimLeaguesForTransactions(context.Background(), db, activities.ClaimLeaguesForTransactionsParams{BatchSize: 10})
-	if err != nil {
-		t.Fatalf("claim: %v", err)
-	}
-	if len(got) != 1 || got[0].LeagueID != "expired-claim" {
-		t.Fatalf("expected only expired-claim to be re-claimable, got %+v", got)
-	}
-}
-
-func TestClaimLeagues_ConcurrentClaimsAreDisjoint(t *testing.T) {
-	db := newPGTestDB(t)
-	now := time.Now().UTC()
-	for i := 0; i < 20; i++ {
-		seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: fmt.Sprintf("lg%02d", i), LastFetchedAt: &now})
-	}
-
-	a := &activities.DataFetchActivities{DB: db}
-	var mu sync.Mutex
-	seen := map[string]int{}
-	var wg sync.WaitGroup
-	for w := 0; w < 2; w++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			got, err := a.ClaimLeaguesForTransactions(context.Background(), db, activities.ClaimLeaguesForTransactionsParams{BatchSize: 10})
-			if err != nil {
-				t.Errorf("claim: %v", err)
-				return
-			}
-			mu.Lock()
-			defer mu.Unlock()
-			for _, s := range got {
-				seen[s.LeagueID]++
-			}
-		}()
-	}
-	wg.Wait()
-	if len(seen) != 20 {
-		t.Errorf("expected 20 distinct leagues claimed, got %d", len(seen))
-	}
-	for id, n := range seen {
-		if n > 1 {
-			t.Errorf("league %s claimed %d times", id, n)
-		}
-	}
-}
-
 func TestClaimLeaguesForDrafts_Eligibility(t *testing.T) {
 	db := newPGTestDB(t)
 	now := time.Now().UTC()
