+++ backend/internal/activities/params.go
@@ -36,30 +36,6 @@ type SyncLeagueDraftsBatchParams struct {
 	Concurrency int
 }
 
-type ClaimLeaguesForTransactionsParams struct {
-	BatchSize int
-}
-
-// LeagueTransactionState carries the league ID, season, and leg cursor for one
-// claimed league, as returned by ClaimLeaguesForTransactions.
-type LeagueTransactionState struct {
-	LeagueID       string
-	Season         string
-	LastLegFetched *int
-}
-
-// LeagueTransactionFetchResult is internal/fdb's fetch result for one
-// claimed league's transaction sync — see FetchLeagueTransactions. Rows are
-// split cloud/archive at fetch time (age-based routing, same rule
-// FetchLeagueTransactions always applied); FlushLeagueTransactions is the
-// batch-write counterpart that persists them.
-type LeagueTransactionFetchResult struct {
-	LeagueID    string
-	CloudRows   []models.SleeperTransaction
-	ArchiveRows []models.SleeperTransaction
-	MaxLegSeen  int // 0 if no new legs were found this run
-}
-
 // SyncBatchResult summarizes one batch activity execution. Failed leagues keep
 // their claim and re-enter the queue when it expires.
 type SyncBatchResult struct {
