+++ backend/internal/transactioncron/claim_pg_test.go
@@ -1,0 +1,151 @@
+package transactioncron_test
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"sync"
+	"testing"
+	"time"
+
+	"gorm.io/gorm"
+
+	"backend/internal/models"
+	"backend/internal/testutil"
+	"backend/internal/transactioncron"
+)
+
+// newPGTestDB opens TEST_DATABASE_URL inside a fresh throwaway schema and
+// migrates SleeperLeague into it. Skips the test when the env var is unset —
+// the claim query uses FOR UPDATE SKIP LOCKED, which SQLite cannot express.
+func newPGTestDB(t *testing.T) *gorm.DB {
+	t.Helper()
+	dsn := os.Getenv("TEST_DATABASE_URL")
+	if dsn == "" {
+		t.Skip("TEST_DATABASE_URL not set; claim tests need Postgres (FOR UPDATE SKIP LOCKED)")
+	}
+	scopedDSN := testutil.NewPGSchema(t, dsn, "transactioncron_claim_test")
+	db := testutil.OpenGORM(t, scopedDSN)
+	if err := db.AutoMigrate(&models.SleeperLeague{}); err != nil {
+		t.Fatalf("automigrate: %v", err)
+	}
+	return db
+}
+
+func seedLeague(t *testing.T, db *gorm.DB, l models.SleeperLeague) {
+	t.Helper()
+	if l.Season == "" {
+		l.Season = "2026"
+	}
+	if l.LeagueType == "" {
+		l.LeagueType = "redraft"
+	}
+	if err := db.Create(&l).Error; err != nil {
+		t.Fatalf("seed league %s: %v", l.SleeperLeagueID, err)
+	}
+}
+
+func TestClaimLeagues_OrderingLimitAndStamp(t *testing.T) {
+	db := newPGTestDB(t)
+	now := time.Now().UTC()
+	old := now.Add(-48 * time.Hour)
+	recent := now.Add(-1 * time.Hour)
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "never", LastFetchedAt: &now})
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "oldest", LastFetchedAt: &now, LastTransactionsFetchedAt: &old})
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "recent", LastFetchedAt: &now, LastTransactionsFetchedAt: &recent})
+
+	got, err := transactioncron.ClaimLeaguesForTransactions(context.Background(), db, transactioncron.ClaimLeaguesForTransactionsParams{BatchSize: 2})
+	if err != nil {
+		t.Fatalf("claim: %v", err)
+	}
+	// Which leagues get claimed is what matters (RETURNING order is not
+	// guaranteed): batch of 2 must be the never-fetched and the oldest.
+	claimed := map[string]bool{}
+	for _, s := range got {
+		claimed[s.LeagueID] = true
+		if s.Season != "2026" {
+			t.Errorf("expected Season populated, got %+v", s)
+		}
+	}
+	if len(got) != 2 || !claimed["never"] || !claimed["oldest"] {
+		t.Fatalf("expected {never, oldest}, got %+v", got)
+	}
+	var stamped int64
+	db.Model(&models.SleeperLeague{}).Where("claimed_at IS NOT NULL").Count(&stamped)
+	if stamped != 2 {
+		t.Errorf("expected 2 rows stamped claimed_at, got %d", stamped)
+	}
+}
+
+func TestClaimLeagues_ExcludesIneligible(t *testing.T) {
+	db := newPGTestDB(t)
+	now := time.Now().UTC()
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "skipped", LastFetchedAt: &now, SkippedAt: &now})
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "unfetched"}) // last_fetched_at NULL
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "old-season", Season: "2024", LastFetchedAt: &now})
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "done-complete", Status: "complete", LastFetchedAt: &now, LastTransactionsFetchedAt: &now})
+	// complete but never transaction-synced: still eligible
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "complete-unsynced", Status: "complete", LastFetchedAt: &now})
+
+	got, err := transactioncron.ClaimLeaguesForTransactions(context.Background(), db, transactioncron.ClaimLeaguesForTransactionsParams{BatchSize: 10})
+	if err != nil {
+		t.Fatalf("claim: %v", err)
+	}
+	if len(got) != 1 || got[0].LeagueID != "complete-unsynced" {
+		t.Fatalf("expected only complete-unsynced, got %+v", got)
+	}
+}
+
+func TestClaimLeagues_RespectsAndExpiresClaims(t *testing.T) {
+	db := newPGTestDB(t)
+	now := time.Now().UTC()
+	fresh := now.Add(-1 * time.Minute)
+	stale := now.Add(-30 * time.Minute)
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "fresh-claim", LastFetchedAt: &now, ClaimedAt: &fresh})
+	seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: "expired-claim", LastFetchedAt: &now, ClaimedAt: &stale})
+
+	got, err := transactioncron.ClaimLeaguesForTransactions(context.Background(), db, transactioncron.ClaimLeaguesForTransactionsParams{BatchSize: 10})
+	if err != nil {
+		t.Fatalf("claim: %v", err)
+	}
+	if len(got) != 1 || got[0].LeagueID != "expired-claim" {
+		t.Fatalf("expected only expired-claim to be re-claimable, got %+v", got)
+	}
+}
+
+func TestClaimLeagues_ConcurrentClaimsAreDisjoint(t *testing.T) {
+	db := newPGTestDB(t)
+	now := time.Now().UTC()
+	for i := 0; i < 20; i++ {
+		seedLeague(t, db, models.SleeperLeague{SleeperLeagueID: fmt.Sprintf("lg%02d", i), LastFetchedAt: &now})
+	}
+
+	var mu sync.Mutex
+	seen := map[string]int{}
+	var wg sync.WaitGroup
+	for w := 0; w < 2; w++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			got, err := transactioncron.ClaimLeaguesForTransactions(context.Background(), db, transactioncron.ClaimLeaguesForTransactionsParams{BatchSize: 10})
+			if err != nil {
+				t.Errorf("claim: %v", err)
+				return
+			}
+			mu.Lock()
+			defer mu.Unlock()
+			for _, s := range got {
+				seen[s.LeagueID]++
+			}
+		}()
+	}
+	wg.Wait()
+	if len(seen) != 20 {
+		t.Errorf("expected 20 distinct leagues claimed, got %d", len(seen))
+	}
+	for id, n := range seen {
+		if n > 1 {
+			t.Errorf("league %s claimed %d times", id, n)
+		}
+	}
+}
