+++ backend/internal/transactioncron/transactioncron.go
@@ -2,11 +2,7 @@
 // stale leagues' transaction backlog and syncs it via internal/fdb.RunPool,
 // which fetches each league's transactions concurrently and batches the
 // resulting DB writes into one bulk statement per flush rather than one
-// write per league. It replaced the Temporal-based transaction-sync pipeline
-// (workflows.TransactionSyncDispatcher / activities.SyncLeagueTransactionsBatch,
-// both since removed) — claim-based resilience via FOR UPDATE SKIP LOCKED and
-// a claimed_at TTL is already database-native, so Temporal's orchestration
-// machinery was redundant for this workload shape.
+// write per league.
 package transactioncron
 
 import (
@@ -25,29 +21,28 @@ import (
 const pollInterval = 200 * time.Millisecond
 
 // Config holds the transaction-sync cron job's tuning knobs, read from env.
-// Keep PoolSize comfortably under DB_MAX_OPEN_CONNS, though the constraint is
-// looser than it used to be: fetch no longer holds a connection across its
-// Sleeper HTTP calls (only flush does, briefly, once per batch), so the
-// binding limit on PoolSize is now Sleeper API concurrency, not DB connection
-// starvation.
+// Keep PoolSize comfortably under DB_MAX_OPEN_CONNS: fetch doesn't hold a
+// connection across its Sleeper HTTP calls (only flush does, briefly, once
+// per batch), so the binding limit on PoolSize is Sleeper API concurrency,
+// not DB connection starvation.
 type Config struct {
 	PoolSize    int // CRON_TXN_POOL_SIZE, default 10
 	RefillBatch int // CRON_TXN_REFILL_BATCH, default 4
 	// BatchSize flushes accumulated league results once this many have been
 	// fetched.
 	BatchSize int // CRON_TXN_BATCH_SIZE, default 20
-	// BatchFlushIntervalSec flushes accumulated results at least this often,
+	// BatchFlushInterval flushes accumulated results at least this often,
 	// even short of BatchSize, so results don't sit indefinitely.
-	BatchFlushIntervalSec int // CRON_TXN_BATCH_FLUSH_INTERVAL_SECONDS, default 5
+	BatchFlushInterval time.Duration // CRON_TXN_BATCH_FLUSH_INTERVAL_DURATION, default 5s
 }
 
 // LoadConfig reads Config from env, clamped to at least 1.
 func LoadConfig() Config {
 	return Config{
-		PoolSize:              max(helpers.GetEnv("CRON_TXN_POOL_SIZE", 10), 1),
-		RefillBatch:           max(helpers.GetEnv("CRON_TXN_REFILL_BATCH", 4), 1),
-		BatchSize:             max(helpers.GetEnv("CRON_TXN_BATCH_SIZE", 20), 1),
-		BatchFlushIntervalSec: max(helpers.GetEnv("CRON_TXN_BATCH_FLUSH_INTERVAL_SECONDS", 5), 1),
+		PoolSize:           max(helpers.GetEnv("CRON_TXN_POOL_SIZE", 10), 1),
+		RefillBatch:        max(helpers.GetEnv("CRON_TXN_REFILL_BATCH", 4), 1),
+		BatchSize:          max(helpers.GetEnv("CRON_TXN_BATCH_SIZE", 20), 1),
+		BatchFlushInterval: helpers.GetEnv("CRON_TXN_BATCH_FLUSH_INTERVAL_DURATION", 5*time.Second),
 	}
 }
 
@@ -77,7 +72,7 @@ type Report struct {
 func RunTransactionSync(ctx context.Context, dfa *activities.DataFetchActivities, cfg Config) (Report, error) {
 	logger := newStdLogger()
 	logger.Info("transaction sync cron starting", "poolSize", cfg.PoolSize, "refillBatch", cfg.RefillBatch,
-		"batchSize", cfg.BatchSize, "batchFlushIntervalSec", cfg.BatchFlushIntervalSec)
+		"batchSize", cfg.BatchSize, "batchFlushInterval", cfg.BatchFlushInterval)
 	start := time.Now()
 
 	state, err := dfa.Sleeper.GetNFLState(ctx)
@@ -92,16 +87,18 @@ func RunTransactionSync(ctx context.Context, dfa *activities.DataFetchActivities
 			RefillBatch:        cfg.RefillBatch,
 			PollInterval:       pollInterval,
 			BatchSize:          cfg.BatchSize,
-			BatchFlushInterval: time.Duration(cfg.BatchFlushIntervalSec) * time.Second,
+			BatchFlushInterval: cfg.BatchFlushInterval,
 		},
-		func(ctx context.Context, db *gorm.DB, n int) ([]activities.LeagueTransactionState, error) {
-			return dfa.ClaimLeaguesForTransactions(ctx, db, activities.ClaimLeaguesForTransactionsParams{BatchSize: n})
+		func(ctx context.Context, db *gorm.DB, n int) ([]LeagueTransactionState, error) {
+			return ClaimLeaguesForTransactions(ctx, db, ClaimLeaguesForTransactionsParams{BatchSize: n})
 		},
-		func(ctx context.Context, db *gorm.DB, lg activities.LeagueTransactionState) (activities.LeagueTransactionFetchResult, error) {
-			return dfa.FetchLeagueTransactions(ctx, db, lg, activities.MaxLegForLeague(lg.Season, state))
+		func(ctx context.Context, db *gorm.DB, lg LeagueTransactionState) (LeagueTransactionFetchResult, error) {
+			return FetchLeagueTransactions(ctx, dfa, lg, MaxLegForLeague(lg.Season, state))
 		},
-		dfa.FlushLeagueTransactions,
-		func(lg activities.LeagueTransactionState, err error, duration time.Duration) {
+		func(ctx context.Context, tx *gorm.DB, batch []LeagueTransactionFetchResult) error {
+			return FlushLeagueTransactions(ctx, dfa, tx, batch)
+		},
+		func(lg LeagueTransactionState, err error, duration time.Duration) {
 			if err != nil {
 				logger.Warn("league transaction sync failed", "leagueID", lg.LeagueID, "error", err, "duration", duration)
 				return
